# Template matching speed