--- a/pyxem/generators/diffraction_generator.py
+++ b/pyxem/generators/diffraction_generator.py
@@ -94,6 +94,8 @@ class DiffractionGenerator(object):
         reciprocal_radius : float
             The maximum radius of the sphere of reciprocal space to sample, in
             reciprocal angstroms.
+        with_direct_beam : bool
+            Include the direct beam peak
 
         Returns
         -------