--- a/pyxem/generators/diffraction_generator.py
+++ b/pyxem/generators/diffraction_generator.py
@@ -95,7 +95,8 @@ class DiffractionGenerator(object):
             The maximum radius of the sphere of reciprocal space to sample, in
             reciprocal angstroms.
         with_direct_beam : bool
-            Include the direct beam peak
+            If True, the direct beam is included in the simulated diffraction
+            pattern. If False, it is not.
 
         Returns
         -------