--- a/pyxem/utils/__init__.py
+++ b/pyxem/utils/__init__.py
@@ -15,4 +15,3 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with pyXem.  If not, see <http://www.gnu.org/licenses/>.
-