---
Template matching efficiency
---

These changes aim to improve the speed of template matching by rewriting the looped calls to `correlate` in `correlate_library` to numpy operations batched over all orientations. In-plane rotations are now included (at user-specified angles/resolution) in the calculation and does not have to be pre-calculated in the diffraction library. This speeds up computation time for the library and reduces storage requirements when sampling many in-plane rotations.

To enable the direct numpy operations in `correlate_library`, some loading code for the library had to be added. `IndexationGenerator.correlate` shuffles the memory in the library dictionary to create separate arrays of the different value types (`orientation`, `pixel_coords`, `intensities`, `pattern_norms`). Only the highest-intensity peaks (currently 100, should be user-specified with a sane default value) are included to ensure non-jagged arrays. The arrays could also be padded to the maximum peak count instead of cut. 

The loading code also calculates all pixel coordinates for in-plane rotations by direct rotation using the user-specified list of angles. The library loading is only done once per call to `IndecationGenerator.correlate` and is therefore relatively quick when matching over several diffraction patterns.

The library loading code keeps the existing library format. Most of this loading code can be simplified (only applying rotation to the pixel coordinates is needed) if the format can be changed. Instead of a format
```
library[phase][orientation] -> {'pixel_coords': [Nx2], 'intensities': [N], ...}
```
(for N peaks) where all data for a single phase and orientation is grouped, the different fields can be separated into separate arrays, only grouped by phase:
```
library[phase] -> {'orientations': [MxNx3], 'pixel_coords': [MxNx2], 'intensities': [MxN], ...}
```
(M orientations) where values at the same index `0 <= m < M` belong together. This works best if all orientations store the same number of peaks, which can be solved by padding (takes more space) or cutting (potentially loosing data in very symmetric directions).

The `peaks_from_best_template` function has also been updated to generate missing entries for matches at in-plane rotations not included in the library. This requires some additional parameters. For convenience, these should maybe be stored in the library.

**Work in progress**
- [x] Decide on library format
- [x] Performance comparison to current implementation, both few and many in-plane rotations
- [x] Store diffraction simulation parameters in the library?