@pc494 2019-03-11
This looks exciting, and I'm happy to review it when it's done.

In the short term:
a) Have you read Wu & Zaefferer from 2009 [1] ? This was how I had imagined getting our speed ups.
b) Which bit of the current code do you think is the slow bit? For example how do the load and the dot product compare?

[1] - https://www.sciencedirect.com/science/article/abs/pii/S0304399109001363?via%3Dihub