@shogas 2019-03-11
a) Thank you, I had not read that article, but seen the concept. The rotation independence of the circular projection is interesting. More than speed, I think this technique could improve the quality of the matching, given good enough pattern simulations. I should probably try implementing a prototype in pyxem, though I am not sure that the current pattern simulation and `as_signal` will give good results for  experimental data.

b) The initial library conversion is negligible for anything more than a few patterns. For a test on 1500 orientations x 360 in-plane, it seems like
```image_intensities = image[pixel_coords[:, :, :, 1], pixel_coords[:, :, :, 0]]```
is about 1/3 and
```correlations = np.sum(image_intensities * intensities, axis=2) / pattern_norms```
2/3 of the total computation time.