@shogas 2019-03-14
I would like some input on the first and third TODO above, as they are more of a design decision:

Changing the library might break existing code that relies on its internal structure and existing pickled libraries cannot be loaded (could special case). On the plus side, it reduces the complexity of the library by removing an extra representation (and some lines of code for conversion). How concerned are we about backward compatibility for the libraries? Also, should we add the parameters used to generate the library in the library itself. That would remove the three new parameters to `plot_best_matching_results_on_signal` (`diffraction_generator`, `reciprocal_radius`, `with_direct_beam`), but I am not sure they belong in the library.

I wrote some performance tests to get a feel for the changes to the correlation function and was mostly interested in the behaviour for smaller libraries. Seems like they are a win also there. The new implementation also scales better with the number of cores, since the numpy calculations releases the GIL. (Still not a proper substitute for navigation-space parallelisation, but that is a task for another time). Note that these tests are not meant to be statistically significant and the absolute numbers are meaningless. They are just to see some relative trends.

Summary of the results:
The timings are for 32 diffraction patterns with a library covering all cubic (and hexagonal for the first four rows) unique directions with 360 (first four) / 1 (last four) in-plane rotations.

| Cores | Templates | Time (s) | Description |
| ----: | --------: | -------: | ----------- |
|     1 |   1009800 |  38.0766 | Single-threaded, new |
|     1 |   1009800 |  304.76  | Single-threaded, current |
|     4 |   1009800 |  13.0912 | Multi-threaded, new |
|     4 |   1009800 |  558.995 | Multi-threaded, current |
|     1 |      1052 | 0.029612 | Single-threaded, new |
|     1 |      1052 | 0.405411 | Single-threaded, current |
|     4 |      1052 | 0.017402 | Multi-threaded, new |
|     4 |      1052 | 0.548826 | Multi-threaded, current |

Relevant line timings for the single-threaded run, with a library of all cubic and hexagonal directions (including in-plane), 1 degree resolution (first row in table above):

New implementation:

```
Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   151                                           def correlate_library_sparse_dot_numpy(image, library_entries, n_largest, mask):
# ...
   155        48        268.0      5.6      0.0          for phase_index, library_entry in enumerate(library_entries):
# ...
   161                                                       # Extract experimental intensities from the diffraction image
   162        32   60993115.0 1906034.8    68.5              image_intensities = image[pixel_coords[:, :, :, 1], pixel_coords[:, :, :, 0]]
   163                                                       # Correlation is the normalized dot product
   164        32   27544101.0 860753.2     30.9              correlations = np.sum(image_intensities * intensities, axis=2) / pattern_norms
# ...

Current implementation:

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   184                                           def correlate(image, pattern_intensities, pixel_coordinates, pattern_normalization):
   185                                               # The x,y choice here is correct. Basically the numpy/hyperspy conversion is a danger
   186  16156800  244563716.0     15.1     77.3      image_intensities = image[pixel_coordinates[:, 1], pixel_coordinates[:, 0]]
   187
   188  16156800   71756031.0      4.4     22.7      return np.dot(image_intensities, pattern_intensities) / pattern_normalization


Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   193                                           def correlate_library_sparse_dot_old(image, library_entries, n_largest, mask):
# ...
   196        48        421.0      8.8      0.0          for phase_index, library_entry in enumerate(library_entries):
# ...
   204     44912      83990.0      1.9      0.0              for i in range(orientations.shape[0]):
   205  16201680   31719429.0      2.0      4.5                  for j in range(pixel_coords.shape[0]):
   206  16156800  510296510.0     31.6     71.6                      correlation = correlate(image, intensities[i], pixel_coords[j][j], pattern_norms[i])
   207  16156800  168956254.0     10.5     23.7                      correlations[j, i, :] = *orientations[i], correlation
# ...
```


The other runs show a similar distribution of timings. Note that I added an extra loop in the current implementation during testing to keep the library size reasonable while testing.