@dnjohnstone 2019-03-15
> Changing the library might break existing code that relies on its internal structure and existing pickled libraries cannot be loaded (could special case). On the plus side, it reduces the complexity of the library by removing an extra representation (and some lines of code for conversion). How concerned are we about backward compatibility for the libraries? Also, should we add the parameters used to generate the library in the library itself. That would remove the three new parameters to `plot_best_matching_results_on_signal` (`diffraction_generator`, `reciprocal_radius`, `with_direct_beam`), but I am not sure they belong in the library.

I wouldn't worry about backward compatibility at the moment. Until there is a proper release of the code it should be expected that there will be things that change without warning.

I'm quite relaxed about storing library generation parameters in the library - can we store it as attributes kind of like "simulation metadata"? If so, I think I quite like that idea as a platform for 

Great about the timings.